code/requirements.txt
pandas==0.24.2
numpy==1.13.3
cdsapi==0.1.4
Shapely==1.6.4.post2
matplotlib==2.1.1
xarray==0.12.1
urllib3==1.22
pytz==2018.3
descartes==1.1.0
geopandas==0.5.0
python_dateutil==2.8.0
netCDF4==1.5.1.2
pyshp==2.1.0
